Add calls to the segmentation depth algorithms in ObjectDetector.py
Add in ObjectDetector.py to call the two depth algorithms at Segmentation.py. For this, include a one-time subscription for the realsense intrinsics. This, effectivily finished the addition of the Intel RealSense Camera and the addition of the depth algorithms; only is missing to include a show_image() utility to visualize in development.